Added preamp module for squaring waves
A pre-amplificator module was added, so that one can use the timecode
signal from some DVS CDs or vinyls.
They are always encoded as amplitude-modulated sine/cos waves, so that it
is always possible to square the waves and obtain a 2-bits incremental
encoder signal, provided that the volume is reasonable.
A red LED will signal to the user that the signal amplitude is not enough
to read quadrature encoder signal, or simpli that the CD/vinyl is paused.

TODO: add a double RCA connector for inputs.

Andrea Zoppi