Window comparator for preamp calibration
The positive voltage reference for opamps is now reduced to a simple
potentiometer, sithout impedance decoupling (not strictly needed).

Two opamps are used as comparators for windowing of the rectified input
signals. This way it is possible to warn the user if the signal is too low
like before (bad comparator results) and also if the signal is too strong,
thus leading to a potentially wring duty cycle of the squared wave.

Andrea Zoppi