Device library behavior and interrupt changes
The device API was changed so that the incremental encoder and the mouse
sensor status are cached into global variables, instead of passing values
by function calls, since there are not multiple instances of such devices.
This improves performance and optimizes memory usage.

It is now possible to choose if the mouse sensor code reacts to interrupts
or polls the status. Since the interrupt is simply cached and just wastes
ISR code, it is a better idea to poll it just before building the HID
report.

Andrea Zoppi