Sensor performance not forced anymore
Even tough it sounds strange, sensor performance is better if not always
forced to the maximum. In fact, when forced it loses accuracy with small
motions, which should be the main objective of the project.
Not forcing performance settings makes the sensor highly reactive to the
smallest movements, and automatically adjusts the performance when the
motion is detected. The built-in algorithm works betetr than expected!

Andrea Zoppi